use a common component for token header amounts
